Fixed Everything My bad

+++ DataTypes/GameObject.cs
@@ -2,7 +2,7 @@
 {
     public  class GameObject
     {
-        private int _gameId = 0;
+        private long _gameId = 0;
         private string _gameName = string.Empty;   
         private string _gameThumbnail = string.Empty;
         private string _gameUrl = string.Empty;
@@ -10,7 +10,7 @@
         /// <summary>
         /// Gets the place ID.
         /// </summary>
-        public int PlaceId { get { return _gameId; } }
+        public long PlaceId { get { return _gameId; } }
 
         /// <summary>
         /// Gets the game name.
@@ -28,7 +28,7 @@
         public string Url { get { return _gameUrl; } }
 
 
-        public GameObject(int gameId, string gameName, string gameThumbnail)
+        public GameObject(long gameId, string gameName, string gameThumbnail)
         {
             _gameId = gameId;
             _gameName = gameName;
